requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
aiohttp>=3.8.1
python-dotenv>=0.19.0
redis>=4.3.4

# Consensus and distributed systems
aioredis>=2.0.1
asyncio>=3.4.3
aiodns>=3.0.0

# Testing
pytest>=7.1.2
pytest-asyncio>=0.18.3
httpx>=0.23.0
locust>=2.8.6

# Monitoring and metrics
prometheus-client>=0.14.1
grafana-api>=1.0.3

# Cache and performance
cachetools>=5.0.0

# Security
cryptography>=37.0.0
pyOpenSSL>=22.0.0

# Documentation
pyyaml>=6.0.0
swagger-ui-bundle>=0.0.9

# Additional tools
grpcio>=1.44.0
protobuf>=3.20.0
kafka-python>=2.0.2  # Optional for message streaming